requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
openai==1.84.0
flask-cors==4.0.0
httpx==0.24.1 